+++ Xext/meson.build
@@ -9,17 +9,12 @@ srcs_xext = [
     'xtest.c',
 ]
 
-hdrs_xext = [
-    'syncsdk.h',
-]
-
 if build_dpms
     srcs_xext += 'dpms.c'
 endif
 
 if build_mitshm
     srcs_xext += 'shm.c'
-    hdrs_xext += ['shmint.h']
 endif
 
 if build_hashtable
@@ -52,7 +47,6 @@ endif
 
 if build_xv
     srcs_xext += ['xvmain.c', 'xvdisp.c', 'xvmc.c']
-    hdrs_xext += ['xvdix.h', 'xvmcext.h']
 endif
 
 libxserver_xext = static_library('xserver_xext',
@@ -60,10 +54,6 @@ libxserver_xext = static_library('xserver_xext',
     include_directories: inc,
     dependencies: common_dep,
 )
-
-if build_xorg
-    install_data(hdrs_xext, install_dir: xorgsdkdir)
-endif
 
 subdir('damage')
 subdir('dri2')
