+++ Xext/shm.c
@@ -48,6 +48,7 @@ in this Software without prior written authorization from The Open Group.
 #include "dix/screen_hooks_priv.h"
 #include "dix/screenint_priv.h"
 #include "dix/window_priv.h"
+#include "include/shmint.h"
 #include "miext/extinit_priv.h"
 #include "os/auth.h"
 #include "os/busfault.h"
@@ -68,7 +69,6 @@ in this Software without prior written authorization from The Open Group.
 #include "gcstruct.h"
 #include "extnsionst.h"
 #include "servermd.h"
-#include "shmint.h"
 #include "xace.h"
 #include "protocol-versions.h"
