+++ Xext/sync.c
@@ -61,6 +61,7 @@ PERFORMANCE OF THIS SOFTWARE.
 #include "dix/dix_priv.h"
 #include "dix/request_priv.h"
 #include "dix/screenint_priv.h"
+#include "include/syncsdk.h"
 #include "miext/extinit_priv.h"
 #include "os/bug_priv.h"
 #include "os/osdep.h"
@@ -72,7 +73,6 @@ PERFORMANCE OF THIS SOFTWARE.
 #include "pixmapstr.h"
 #include "resource.h"
 #include "syncsrv.h"
-#include "syncsdk.h"
 #include "protocol-versions.h"
 #include "inputstr.h"
 #include "misync_priv.h"
