+++ hw/kdrive/src/kxv.h
@@ -38,6 +38,8 @@ of the copyright holder.
 #ifndef _XVDIX_H_
 #define _XVDIX_H_
 
+#include "include/xvdix.h"
+
 #include "scrnintstr.h"
 #include "regionstr.h"
 #include "windowstr.h"
@@ -46,8 +48,6 @@ of the copyright holder.
 #include "resource.h"
 #include "gcstruct.h"
 #include "dixstruct.h"
-
-#include "../../Xext/xvdix.h"
 
 #define VIDEO_OVERLAID_IMAGES			0x00000004
 #define VIDEO_OVERLAID_STILLS			0x00000008
