+++ include/shmint.h
@@ -23,6 +23,7 @@
 #ifndef _SHMINT_H_
 #define _SHMINT_H_
 
+#include <X11/Xmd.h>
 #include <X11/extensions/shmproto.h>
 
 #include "screenint.h"
