+++ present/present_priv.h
@@ -27,13 +27,13 @@
 #include <X11/Xmd.h>
 
 #include "include/present.h"
+#include "include/syncsdk.h"
 
 #include "scrnintstr.h"
 #include "misc.h"
 #include "list.h"
 #include "windowstr.h"
 #include "dixstruct.h"
-#include <syncsdk.h>
 #include <syncsrv.h>
 #include <xfixes.h>
 #include <randrstr.h>
